+++ AppKit/NSMenu.subproj/NSMenuWindow.h
@@ -15,7 +15,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 }
 
 - initWithMenu:(NSMenu *)menu;
-- initWithMenu:(NSMenu *)menu overflowAtIndex:(unsigned)overflowIndex;
+- initWithMenu:(NSMenu *)menu overflowAtIndex:(NSUInteger)overflowIndex;
 
 - (NSMenuView *)menuView;
