+++ AppKit/NSMenu.subproj/NSOverflowMenuView.m
@@ -11,7 +11,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 @implementation NSOverflowMenuView
 
--initWithMenu:(NSMenu *)menu overflowAtIndex:(unsigned)overflowIndex {
+-initWithMenu:(NSMenu *)menu overflowAtIndex:(NSUInteger)overflowIndex {
    _overflowIndex=overflowIndex;
 
    return [self initWithMenu:menu];
